read_dir_tostring -> create_autoindex_body
