+++ srcs/utils/file_io_utils.cpp
@@ -1,6 +1,5 @@
 #include "utils/file_io_utils.hpp"
 
-#include <dirent.h>
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <unistd.h>
@@ -40,29 +39,6 @@ bool is_dir_exists(const std::string &path) {
   } else {
     return false;
   }
-}
-
-// directory内のファイル・ディレクトリ名をstringに格納して返す
-// TODO: htmlに変換する
-std::string read_dir_tostring(const std::string &file_path) {
-  DIR                     *dir = xopendir(file_path.c_str());
-  struct dirent           *diread;
-  std::vector<std::string> files;
-
-  // xopendir()が成功している時点でdirは有効なものなので
-  // 他のシステムコールはエラーを起こさないことが前提の実装
-  while ((diread = readdir(dir)) != NULL) {
-    files.push_back(diread->d_name);
-  }
-  closedir(dir);
-
-  std::string                        ret;
-  std::vector<std::string>::iterator it = files.begin();
-  for (; it != files.end(); it++) {
-    ret += " " + *it;
-  }
-
-  return ret;
 }
 
 std::string read_file_tostring(const std::string &path) {
